+++ js/similar-ads.js
@@ -69,7 +69,6 @@ const renderSimilarList = (similarAds) => {
         .bindPopup(
           popup,
         );
-
     });
 };
